Reorganise folders / filenames Refactor search slightly to use profiles
