+++ prez/models/catprez_item.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel, root_validator
 from rdflib import URIRef
 from rdflib.namespace import DCTERMS, XSD, DCAT, Namespace
 
-from prez.services.sparql_utils import sparql_query_non_async
+from prez.sparql.methods import sparql_query_non_async
 
 PREZ = Namespace("https://prez.dev/")
