+++ prez/reference_data/prez_ns.py
@@ -1,3 +1,4 @@
 from rdflib import Namespace
 
 PREZ = Namespace("https://prez.dev/")
+ALTREXT = Namespace("http://www.w3.org/ns/dx/conneg/altr-ext#")
