+++ prez/routers/catprez.py
@@ -8,7 +8,7 @@ from prez.models.catprez_item import CatalogItem
 from prez.models.catprez_listings import CatalogMembers
 from prez.models.profiles_and_mediatypes import ProfilesMediatypesInfo
 from prez.renderers.renderer import return_from_queries, return_profiles
-from prez.services.sparql_queries import (
+from prez.sparql.objects_listings import (
     generate_listing_construct_from_uri,
     generate_listing_count_construct,
     generate_item_construct,
