+++ prez/routers/vocprez.py
@@ -1,16 +1,15 @@
 import logging
-import time
 from typing import Optional
 
 from fastapi import APIRouter, Request
-from rdflib import SKOS, URIRef
+from rdflib import URIRef
 from starlette.responses import PlainTextResponse
 
 from prez.models.profiles_and_mediatypes import ProfilesMediatypesInfo
 from prez.models.vocprez_item import VocabItem
 from prez.models.vocprez_listings import VocabMembers
 from prez.renderers.renderer import return_from_queries, return_profiles
-from prez.services.sparql_queries import (
+from prez.sparql.objects_listings import (
     generate_listing_construct_from_uri,
     generate_listing_count_construct,
     generate_item_construct,
