+++ prez/sparql/methods.py
